added new file with new funcitons
