+++ views/dcMovement.js
@@ -1,6 +1,7 @@
 var map = require('../views/mapCreate.js');
 var carCreation = require('./carCreation.js');
 var general = require('../views/general.js');
+var carPositioning = require('./carPositioning.js');
 
 carCreation.createDumbCars();
 var carArray = carCreation.getCarArr();
@@ -104,27 +105,6 @@ function collisionAvoidanceCheck(carId) {
   return shortestDistance;
 }
 
-// Returns the edgeId of the passed in cars next edge on it's current route
-function getNextEdgeInRoute(carId) {
-  var edgeArray = map.getEdgeArray();
-  var currentCar = carCreation.getCar(carId);
-  var currentEdgeEnd = map.getEdgeObject(currentCar._currentEdgeId).endNodeId;
-  var nextEdgeStart = currentEdgeEnd;
-
-  // Finds the ID of the next node in the route
-  if (nextEdgeStart != currentCar.route[currentCar.route.length - 1]) {
-    var nextEdgeEnd = currentCar.route[currentCar.route.indexOf(nextEdgeStart) + 1];
-
-    // Scan through all edges to find the next one on the route
-    for (var i = 1; i < edgeArray.length; i++) {
-      // Switch to this edge
-      if (edgeArray[i].startNodeId == nextEdgeStart && edgeArray[i].endNodeId == nextEdgeEnd) {
-        return edgeArray[i].edgeId;
-      }
-    }
-  }
-}
-
 // This moves the current car onto the next edge in its route
 function switchEdge(carId) {
   var currentCar = carCreation.getCar(carId);
@@ -138,7 +118,7 @@ function switchEdge(carId) {
 
   map.removeCarFromEdge(currentCar.carId, currentCar._currentEdgeId, 1); // TODO Will have to update "0"
   //console.log(map.getEdgeObject(currentCar._currentEdgeId));
-  currentCar._currentEdgeId = getNextEdgeInRoute(carId);
+  currentCar._currentEdgeId = carPositioning.getNextEdgeInRoute(carId);
   map.insertCarToEdge(currentCar.carId, currentCar._currentEdgeId, 1); // TODO Will have to update "0"
 
 
@@ -148,7 +128,7 @@ function switchEdge(carId) {
 // Function to check if the next edge in the current vehicles path has space available to enter
 function isRoadBlocked(carId) {
   var currentCar = carCreation.getCar(carId);
-  var nextEdgeId = getNextEdgeInRoute(carId);
+  var nextEdgeId = carPositioning.getNextEdgeInRoute(carId);
   var carsOnNextEdge = map.getCarsOnEdge(nextEdgeId, currentCar._currentLane);
   var nextEdgeStartNode = map.getStartNode(nextEdgeId);
   var nextEdgeStartNodeX = nextEdgeStartNode.x;
@@ -209,59 +189,6 @@ function moveX(xPos, xDestination, speed) {
   return xPos;
 }
 
-function checkIfLaneChangeIsNeeded(currentLane, currentEdgeId, nextEdgeId) {
-  var currentEdge = map.getEdgeObject(currentEdgeId);
-  var nextEdge = map.getEdgeObject(nextEdgeId);
-  var nextTurn;
-  if (nextEdge) {
-    var currentEdgeOrientation = currentEdge._orientation;
-    var nextEdgeOrientation = nextEdge._orientation;
-    console.log("currentEdgeOrientation " + currentEdgeOrientation + " nextEdgeOrientation " + nextEdgeOrientation);
-
-    if (currentEdgeOrientation != nextEdgeOrientation) {
-      if (currentEdgeOrientation == 0) {
-        if (nextEdgeOrientation == 90) {
-          console.log("left turn");
-          nextTurn = 1;
-        } else {
-          console.log("right turn");
-          nextTurn = 0;
-        }
-      } else if (currentEdgeOrientation == 90) {
-        if (nextEdgeOrientation == 180) {
-          console.log("left turn");
-          nextTurn = 1;
-        } else {
-          console.log("right turn");
-          nextTurn = 0;
-        }
-      } else if (currentEdgeOrientation == 180) {
-        if (nextEdgeOrientation == 270) {
-          console.log("left turn");
-          nextTurn = 1;
-        } else {
-          console.log("right turn");
-          nextTurn = 0;
-        }
-      } else if (currentEdgeOrientation == 270) {
-        if (nextEdgeOrientation == 0) {
-          console.log("left turn");
-          nextTurn = 1;
-        } else {
-          console.log("right turn");
-          nextTurn = 0;
-        }
-      }
-      if (nextTurn == currentLane) {
-        return -1;
-      } else {
-        return nextTurn;
-      }
-    }
-  }
-  return false;
-}
-
 function moveCar(carInfo) {
   // Get car information from the object
   var carId = carInfo.carId;
@@ -312,8 +239,8 @@ function moveCar(carInfo) {
     if (xDifference <= 0 && yDifference <= 0) {
       switchEdge(carInfo.carId);
 
-      var nextEdgeId = getNextEdgeInRoute(carId);
-      var needToChangeLane = checkIfLaneChangeIsNeeded(carInfo._currentLane, carInfo._currentEdgeId, nextEdgeId);
+      var nextEdgeId = carPositioning.getNextEdgeInRoute(carId);
+      var needToChangeLane = carPositioning.checkIfLaneChangeIsNeeded(carInfo._currentLane, carInfo._currentEdgeId, nextEdgeId);
 
       carInfo._shouldChangeLane= needToChangeLane;
       console.log(carInfo._shouldChangeLane);
